fix: 요구 사항에 맞게 코드 수정
