--- src/app/main/Main.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import { MainLayout } from './Main.styles';
-import Todo from '../todo/Todo';
-
-const Main = () => {
-  return (
-    <MainLayout>
-      <Todo />
-    </MainLayout>
-  );
-};
-
-export default Main;
