--- src/app/main/Main.styles.jsx
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-export const MainLayout = styled.main`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100vw;
-  height: 100vh;
-  background-color: #bdbfc1;
-`;
