+++ src/app/todo/TodoTemplate.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { TodoContainer } from './TodoTemplate.styles';
+
+const TodoTemplate = ({ children }) => {
+  return <TodoContainer>{children}</TodoContainer>;
+};
+
+export default TodoTemplate;
